Change structure to better
Use shared pointer (vector change adress)
Use vectors

+++ test/Triangel.cpp
@@ -8,17 +8,17 @@ void Triangel2Verticies(std::vector<Triangel> &Triangels, std::vector<verticies>
 
 		for (int i = 0; i < 3; i++)
 		{
-			Verticies[i].vertexBufferData.emplace_back((GLfloat)Triangels[i].EdgePos[i*2].x);
-			Verticies[i].vertexBufferData.emplace_back((GLfloat)Triangels[i].EdgePos[i*2 + 1].y);
+			Verticies[i].vertexBufferData.emplace_back(Triangels[i].EdgePos[i*2].x);
+			Verticies[i].vertexBufferData.emplace_back(Triangels[i].EdgePos[i*2 + 1].y);
 		}
 
 		Verticies[i].colorBufferData.clear();
 		
 		for (int i = 0; i < 3; i++)
 		{
-			Verticies[i].vertexBufferData.emplace_back((GLfloat)Triangels[i].EdgeColor[i*3].r);
-			Verticies[i].vertexBufferData.emplace_back((GLfloat)Triangels[i].EdgeColor[i*3 + 1].g);
-			Verticies[i].vertexBufferData.emplace_back((GLfloat)Triangels[i].EdgeColor[i*3 + 2].b);
+			Verticies[i].vertexBufferData.emplace_back(Triangels[i].EdgeColor[i*3].r);
+			Verticies[i].vertexBufferData.emplace_back(Triangels[i].EdgeColor[i*3 + 1].g);
+			Verticies[i].vertexBufferData.emplace_back(Triangels[i].EdgeColor[i*3 + 2].b);
 		}
 	}
 }
